sqs_build/requirements.txt

boto3
pinecone
langchain
fastapi
mangum
PyPDF2
pypdf
pytesseract
Pillow
pandas
langchain-google-genai
langchain-community
# uvicorn[standard]
python-multipart
numpy>=1.22,<2.1
  